refactor(album): consolidate hook return types and improve type imports for better clarity and organization
